名前等のリファクタリング
