+++ configs/settings/cifar10-org-mixer.yaml
@@ -30,8 +30,8 @@ num_epochs: 24
 num_warmup_epochs: 4
 validate_every: 1
 stop_iteration:
-bucket_name: sscr-mlp
-project_name: sscr-mlp
+bucket_name: raft-mlp
+project_name: raft-mlp
 task_name: cifar10-org-mixer
 backend:
 resume_from:
