+++ configs/settings/imagenet-raft-mlp.yaml
@@ -5,8 +5,8 @@ data_path: /datasets/imagenet
 token_mixing_type: ser_pm
 in_channels: 3
 layers:
-  - depth: 36
-    dim: 384
+  - depth: 12
+    dim: 768
     patch_size: 16
     raft_size: 2
 token_expansion_factor: 2
@@ -31,9 +31,9 @@ num_epochs: 305
 num_warmup_epochs: 5
 validate_every: 1
 stop_iteration:
-bucket_name: sscr-mlp
-project_name: sscr-mlp
-task_name: imagenet-sscr-mlp-deep
+bucket_name: raft-mlp
+project_name: raft-mlp
+task_name: imagenet-sscr-mlp
 backend:
 resume_from:
 nproc_per_node:
