+++ libs/consts.py
@@ -3,7 +3,7 @@ CIFAR100 = "cifar100"
 IMAGENET = "imagenet"
 PATCH_SIZE = "patch_size"
 RAFT_SIZE = "raft_size"
-LOGGER_NAME = "SSCR-MLP"
+LOGGER_NAME = "RaftMLP"
 DIM = "dim"
 DEPTH = "depth"
 SER_PM = "ser_pm"
