+++ summary.py
@@ -6,7 +6,7 @@ from ptflops import get_model_complexity_info
 
 from libs.consts import IMAGENET, CIFAR10, CIFAR100
 from libs.datasets import ImageNetGetter, CIFAR10Getter, CIFAR100Getter
-from libs.models import SSCRMLP
+from libs.models import RaftMLP
 
 
 @hydra.main(config_path="configs", config_name="config")
@@ -21,7 +21,7 @@ def run_summary(params: DictConfig) -> None:
         dg = CIFAR100Getter()
     else:
         raise ValueError("Invalid dataset name")
-    model = SSCRMLP(
+    model = RaftMLP(
         layers=params.settings.layers,
         in_channels=dg.channels,
         image_size=dg.image_size,
